Support query by condition in JS Query apis

+++ base/event_store/dao/doc_query.cpp
@@ -116,7 +116,7 @@ bool DocQuery::IsContainExtraConds(EventRaw::DecodedEvent& decodedEvent) const
             if (int64_t intValue = 0; param->AsInt64(intValue)) {
                 paramValue = intValue;
             } else if (uint64_t uintValue = 0; param->AsUint64(uintValue)) {
-                paramValue = static_cast<int64_t>(uintValue);
+                paramValue = uintValue;
             } else if (double dValue = 0; param->AsDouble(dValue)) {
                 paramValue = dValue;
             } else if (std::string sValue; param->AsString(sValue)) {
